Make mypy happy by splitting Array and LogicArray into const
Array and LogicArray had to be invariant since they are mutable.
However this had some issues with Self types not matching because the
bound used the invariant type. Also this limits the types on LogicArray
logical operators to the LogicArrayBase subclasses, since only they are
required to be virtual subclasses at runtime.